user auth finished sign ups work need to mess with login to show welcome prompt
